MyApi/requirements.txt
flask
flask-smorest
flask-sqlalchemy
flask_migrate
gunicorn
requests
flask_cors
